requirements.txt
urllib3==1.10.2
hammertime-http==0.2.2
easyinject==0.3